Refactor Overpass client query building and parsing
